+++ orix/quaternion/orientation.py
@@ -562,7 +562,7 @@ class Orientation(Misorientation):
         dot_outer
         """
         symmetry = self.symmetry.outer(other.symmetry).unique()
-        misorientation = (~self) * other
+        misorientation = other * ~self
         all_dot_products = Rotation(misorientation).dot_outer(symmetry).data
         highest_dot_product = np.max(all_dot_products, axis=-1)
         return Scalar(highest_dot_product)
@@ -577,7 +577,7 @@ class Orientation(Misorientation):
         dot
         """
         symmetry = self.symmetry.outer(other.symmetry).unique()
-        misorientation = (~self).outer(other)
+        misorientation = other.outer(~self)
         all_dot_products = Rotation(misorientation).dot_outer(symmetry).data
         highest_dot_product = np.max(all_dot_products, axis=-1)
         return Scalar(highest_dot_product)
@@ -905,7 +905,7 @@ class Orientation(Misorientation):
         `dp = dparr.compute()`.
         """
         symmetry = self.symmetry.outer(other.symmetry).unique()
-        misorientation = (~self)._outer_dask(other, chunk_size=chunk_size)
+        misorientation = other._outer_dask(~self, chunk_size=chunk_size)
 
         # Summation subscripts
         str1 = "abcdefghijklmnopqrstuvwxy"[: misorientation.ndim]