+++ CHANGELOG.rst
@@ -14,6 +14,11 @@ Added
 -----
 - Python 3.10 support.
 
+Fixed
+-----
+- Orientation disorientation angles and dot products returned from `angle_with()` and
+  `dot()` and `dot_outer()`, which now calculates the misorientation as other * ~self.
+
 2021-12-21 - version 0.8.0
 ==========================
 
@@ -78,7 +83,6 @@ Fixed
 - `CrystalMap.get_map_data()` can return an array of shape (3,) if there are that many
   points in the map.
 - Reading of point groups with "-" sign, like -43m, from EMsoft h5ebsd files.
-
 
 2021-09-07 - version 0.7.0
 ==========================