+++ CHANGELOG.rst
@@ -17,7 +17,8 @@ Added
 Fixed
 -----
 - `Orientation` disorientation angles and dot products returned from `angle_with()` and
-  `dot()` and `dot_outer()`, which now calculates the misorientation as other * ~self.
+  `dot()` and `dot_outer()`, which now calculates the misorientation as `other * ~self`.
+  Disorientation angles `(o2 - o1).angle` and `o1.angle_with(o2)` are now the same.
 
 2021-12-21 - version 0.8.0
 ==========================